fix: address code review issues - random_int, PDO error codes, timestamp consistency, singleTap reset, XP cap
Agent-Logs-Url: https://github.com/codenimar/moonbase/sessions/00a46730-c63b-41e2-bd9b-017752ff2e87

+++ game.php
@@ -182,7 +182,7 @@ if (!$authed) {
 
 <script>
 // Bridge fee constant exposed to JS
-const MOONCOIN_BRIDGE_FEE = <?= (int)MOONCOIN_BRIDGE_FEE_PCT ?>;
+const MOONCOIN_BRIDGE_FEE = <?= (float)MOONCOIN_BRIDGE_FEE_PCT ?>;
 
 // ── Bootstrap ──────────────────────────────────────────────────────────────
 (async () => {
